+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ##### Bug Fixes
 
-- None.
+- Fix false positive when protocols requirements are satisfied in another file from the one that declares the conformance.
 
 ## 2.8.3 (2021-11-29)